Update supported_devices.lua
I had problems getting both my Launchpad Pro Mk 3 and my Linnstrument recognised, but adding these additional lines included helped. It seems that they are case sensitive, and that my Mk 3 required a number to be added on the end, due to the way that Norns Shield is recognising the devices.